were added: 1) Fragment dialogs for creation and deleting words and dictionaries 2) Fragment dialog view 3) Other minor changes in classes and resources
other:
1) DictionaryCreationActivity was deleted

+++ app/src/main/java/by/ctefi/wordbox/database/dao/WordDictionaryJoinDao.kt
@@ -18,7 +18,7 @@ interface WordDictionaryJoinDao {
                 "WHERE word_dictionary.dictionaryId = :dictionaryId " +
                 "ORDER BY word.id DESC"
     )
-    fun getWordsForDictionary(dictionaryId: Int): LiveData<List<Word>>
+    fun getWordsForDictionary(dictionaryId: Long): LiveData<List<Word>>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     fun insertWordForDictionary(wordDictionaryJoin: WordDictionaryJoin)
