+++ app/src/main/java/by/ctefi/wordbox/entity/Word.kt
@@ -5,9 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "word")
 data class Word(
-
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     val original: String,
     val translation: String,
     val meaning: String = ""
